теперь в структуре symbol содержится значение для глобальных констант; пофиксили в процессе баг с поиском инициализированных double; заменили в коде все __int64 -> long long.
git-svn-id: file:///home/cppg/dev/svn/compilers2/trunk@782 a9b74587-2ee6-7842-8a07-03ad48b783cd

+++ cc4x86/src/src/x86_text_output.c
@@ -101,7 +101,7 @@ void text_output_declare_initialized_dword(symbol *sym, long value)
     _out_fmt("public\t_%s\n", sym->sym_name);
 }
 
-void text_output_declare_initialized_qword(symbol *sym, __int64 value)
+void text_output_declare_initialized_qword(symbol *sym, long long value)
 {
     _out_fmt("_%s\tdq\t0%I64xh\n", sym->sym_name, value);
     _out_fmt("public\t_%s\n", sym->sym_name);
